modified:   ZXCPMsrc/zx_display.c 	modified:   ZXCPMsrc/zx_display.h 	modified:   ZXCPMsrc/zx_system.c

+++ ZXCPMsrc/zx_display.h
@@ -25,7 +25,7 @@
 
 extern uint32_t zxlnum;
 extern uint32_t zx_newline_flag;
-extern uint32_t zx_border_color;
+extern uint8_t zx_border_color;
 
 void zxdisp_init();
 void zxdisp_deinit();
